bring logic closer to their component
