+++ frontend/app/src/components/form/node-with-profile-form.tsx
@@ -1,10 +1,9 @@
-import { useAtomValue } from "jotai/index";
-import { genericsState, profilesAtom, schemaState } from "@/state/atoms/schema.atom";
 import { useState } from "react";
 import NoDataFound from "@/screens/errors/no-data-found";
 import { ProfilesSelector } from "@/components/form/profiles-selector";
 import { ObjectFormProps, ProfileData } from "@/components/form/object-form";
 import { NodeForm } from "@/components/form/node-form";
+import { useSchema } from "@/hooks/useSchema";
 
 export const NodeWithProfileForm = ({
   isFilterForm,
@@ -12,31 +11,24 @@ export const NodeWithProfileForm = ({
   currentProfiles,
   ...props
 }: ObjectFormProps) => {
-  const nodes = useAtomValue(schemaState);
-  const generics = useAtomValue(genericsState);
-  const profiles = useAtomValue(profilesAtom);
-
+  const { schema } = useSchema(kind);
   const [selectedProfiles, setSelectedProfiles] = useState<ProfileData[] | undefined>();
 
-  const nodeSchema = [...nodes, ...generics, ...profiles].find((node) => node.kind === kind);
-
-  if (!nodeSchema) {
+  if (!schema) {
     return <NoDataFound message={`${kind} schema not found. Try to reload the page.`} />;
   }
 
   return (
     <>
-      {!isFilterForm && "generate_profile" in nodeSchema && nodeSchema.generate_profile && (
-        <ProfilesSelector
-          schema={nodeSchema}
-          defaultValue={currentProfiles}
-          value={selectedProfiles}
-          onChange={setSelectedProfiles}
-          currentProfiles={currentProfiles}
-        />
-      )}
+      <ProfilesSelector
+        schema={schema}
+        defaultValue={currentProfiles}
+        value={selectedProfiles}
+        onChange={setSelectedProfiles}
+      />
+
       <NodeForm
-        schema={nodeSchema}
+        schema={schema}
         isFilterForm={isFilterForm}
         profiles={selectedProfiles}
         {...props}
